Solution: Replaced file copy entries in csproj with wildcard variant
Adding a new file each day creates unnecessary work and is a potential source of mistakes, if I should forget adding an input file to the list of files to copy on build.
Now with the wildcard entry, each file which fits the pattern will be copied on build automatically and I don't need to enable this manually for each new file.